+++ src/main/java/com/amigoscode/customer/CustomerRepository.java
@@ -2,10 +2,10 @@ package com.amigoscode.customer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 
     boolean existsCustomerByEmail(String email);
 
-    boolean existsCustomerById(Integer id);
+    boolean existsCustomerById(Long id);
 
 }
